[Fix] 코드리뷰 적용 (#285)
DRTextField 관련 enum 분리
